+++ fortpy/testing/executable.py
@@ -287,7 +287,7 @@ info:
 	date >> $(LOG)
 	echo "------------------------------------------------------"| tee -a $(LOG)
 	echo "                     FORTPY"                           | tee -a $(LOG)
-	echo "               >>> version 1.0 <<<                    "| tee -a $(LOG)         
+	echo "               >>> version 1.1 <<<                    "| tee -a $(LOG)         
 	echo "------------------------------------------------------"| tee -a $(LOG)
 	echo -e "Compiling on system  : $(UNAME)"                    | tee -a $(LOG)
 	echo -e "             machine : $(HOSTNAME)"                 | tee -a $(LOG)