+++ tests/isense/original/monte_orderparam.f90
@@ -314,5 +314,6 @@ contains
     call pysave(fir
     call rand
     call this%
+    this%
   end subroutine tracker_finalize
 end module monte_orderparam