+++ tests/parsing.py
@@ -29,6 +29,12 @@ def parse():
             else:
                 print moduledat
 
+        for progdat in c._programfiles[fname]:
+            if args["verbose"]:
+                print c.programs[progdat]
+            else:
+                print progdat
+
 #Create a parser so that the script can receive arguments
 parser = argparse.ArgumentParser(description="Fortpy File Parsing Unit Testing Tool")